Refactor words into its own module
Also fixed missing import in stop_words and core
